+++ backend/src/main/java/com/earntogether/qlysotietkiem/model/DepositSlipModel.java
@@ -2,11 +2,11 @@ package com.earntogether.qlysotietkiem.model;
 
 import lombok.Builder;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Builder
 public record DepositSlipModel(String id, int customerCode, int passbookCode,
                                int type, LocalDate depositDate,
-                               BigInteger money
+                               BigDecimal money
 ) {}
