+++ backend/src/main/java/com/earntogether/qlysotietkiem/model/PassbookModel.java
@@ -1,7 +1,7 @@
 package com.earntogether.qlysotietkiem.model;
  
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public record PassbookModel(int passbookCode, int type, String customerName,
-                            BigInteger money
+                            BigDecimal money
 ){}
